requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-google-genai
chromadb
wikipedia
python-dotenv